New dynamic category page with Next 13
